move the gardenlet chart test to the managed seed directory under /charttest
