+++ pkg/gardenlet/controller/managedseed/charttest/gardenlet_chart.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package chart
+package charttest
 
 import (
 	"context"
